Role fixes

+++ Valour/Server/Services/PlanetRoleService.cs
@@ -52,7 +52,7 @@ public class PlanetRoleService
         if (!_hexColorRegex.IsMatch(role.Color))
             return new TaskResult<PlanetRole>(false, "Invalid hex color");
 
-        role.Position = await _db.PlanetRoles.CountAsync(x => x.PlanetId == role.PlanetId && !x.IsDefault);
+        role.Position = (uint)await _db.PlanetRoles.CountAsync(x => x.PlanetId == role.PlanetId && !x.IsDefault);
         role.Id = IdManager.Generate();
         
         // The index should be the next free index
